requirements.txt
python-dotenv
langchain 
langchain-community 
langchain-openai
langchain-ollama
pypdf
groq
langchain-groq
beautifulsoup4
streamlit
langchain_google_genai
PyPDF2
#faiss-gpu